feat: Split top bar files
